Cleaned version for production

+++ app/Http/Controllers/ReviewController.php
@@ -11,10 +11,11 @@ use App\Services\SteamService;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Mail;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ReviewController extends Controller
 {
-
+    use AuthorizesRequests;
     public function index()
     {
         $sort = request('sort', 'desc');
@@ -37,18 +38,17 @@ class ReviewController extends Controller
     {
         $game_id = $request->query('game_id');
         $steam_id = $request->query('steam_id');
-        // dd($request);
+
         return view('reviews.create', ['game_id' => $game_id, 'steam_id' => $steam_id, 'tags' => Tag::all()]);
     }
 
     public function show(Review $review)
     {
         $review->load('game');
-        // dd($review->game);
+
         if (isset($review->steam_id)) {
-            // $steamService = new SteamService($review->steam_id);
+
             $image_path = $review->game->image_path;
-            // dd($image_path);
         };
         $user = $review->user;
         return view('reviews.show', ['review' => $review, 'image_path' => $image_path, 'user' => $user]);
@@ -90,9 +90,9 @@ class ReviewController extends Controller
         if (request()->has('tag')) {
             $review->tags()->attach(request()->tag);
         }
-        Mail::to($review->user)->queue(
-            new ReviewPosted($review)
-        );
+        // Mail::to($review->user)->queue(
+        //     new ReviewPosted($review)
+        // );
 
 
         return redirect('/reviews');
@@ -105,19 +105,15 @@ class ReviewController extends Controller
 
     public function update(Review $review)
     {
-        // authorize...
+        $this->authorize('edit', $review);
+
         // validation
         request()->validate([
             'title' => ['required', 'min:3', 'max:200'],
             'content' => ['required', 'string', 'max:5000'],
             'score' => ['required'],
         ]);
-        // update and persist
-        // $review = Review::findOrFail($id);
 
-        // $review->title = request('title');
-        // $review->content = request('content');
-        // $review->save();
 
         $review->update([
             'title' => request('title'),
@@ -130,13 +126,10 @@ class ReviewController extends Controller
 
     public function destroy(Review $review)
     {
-        // authorize
-        // delete
-        // Review::findOrFail($id)->delete(); --same as below
-        // $review = Review::findOrFail($id);
+        $this->authorize('edit', $review);
+
         $review->delete();
 
         return redirect('/reviews');
-        // redirect
     }
 }
